requirements.txt
pandas==1.5.3
scikit-learn==1.1.3
psycopg2-binary==2.9.3
